+++ lib/braintree/credit_card_verification.rb
@@ -26,6 +26,7 @@ module Braintree
     attr_reader :merchant_account_id
     attr_reader :processor_response_code
     attr_reader :processor_response_text
+    attr_reader :processor_response_type
     attr_reader :risk_data
     attr_reader :status