+++ lib/braintree/test/transaction_amounts.rb
@@ -5,6 +5,7 @@ module Braintree
     module TransactionAmounts
       Authorize = "1000.00"
       Decline = "2000.00"
+      HardDecline = "2015.00"
       Fail = "3000.00"
     end
   end