+++ spec/integration/braintree/payment_method_spec.rb
@@ -204,18 +204,6 @@ describe Braintree::PaymentMethod do
       venmo_account.customer_id.should == customer.id
     end
 
-    it "raise server error when we attempt to create europe bank account payment method" do
-      customer = Braintree::Customer.create.customer
-      token = SecureRandom.hex(16)
-      expect do
-        Braintree::PaymentMethod.create(
-          :payment_method_nonce => Braintree::Test::Nonce::Europe,
-          :customer_id => customer.id,
-          :token => token
-        )
-      end.to raise_error(Braintree::ServerError)
-    end
-
     it "allows passing the make_default option alongside the nonce" do
       customer = Braintree::Customer.create!
       result = Braintree::CreditCard.create(