+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.92.0
+* Add `processor_response_type` to `Transaction`, `AuthorizationAdjustment`, and `CreditCardVerification`.
+* Add `authorization_expires_at` to `Transaction`
+
 # 2.91.0
 * Add local payment webhook support
 * Add subscription charged unsuccessfully sample webhook to webhook testing gateway